Use the data partition; rootfs is read-only by default; Use devtmpfs instead of mdev.

+++ tools/ota_pre/run.sh
@@ -5,7 +5,13 @@ mv mbrush-fw/mb_ser /root/mb_ser_new
 mv mbrush-fw/mb_conf /root/mb_conf_new
 mv mbrush-fw/mb_etc /root/mb_etc_new
 
-dd if=mbrush-fw/uImage of=/dev/mtdblock1 bs=4k
+if [ -x "$(command -v flash_erase)" ] && [ -x "$(command -v nandwrite)" ]; then
+    flash_erase /dev/mtd1 0 32
+    nandwrite -m -p /dev/mtd1 mbrush-fw/uImage
+else
+    dd if=mbrush-fw/uImage of=/dev/mtdblock1 bs=4k
+fi
+
 rm -rf mbrush-fw
 
 cd /root
@@ -15,7 +21,7 @@ mv mb_ser_new mb_ser
 mv mb_conf_new mb_conf
 mv mb_etc_new mb_etc
 
-ln -sf /root/mb_etc/S80mb /etc/init.d/S80mb
+ln -sf /root/mb_etc/S80mb /etc/init.d/S80mb # skip read only filesystem
 
 sync
 nohup sh -c "sleep 4 && reboot" > /dev/null &
